reorganise blockevents
