+++ src/main/java/io/th0rgal/oraxen/api/events/noteblock/OraxenNoteBlockInteractEvent.java
@@ -1,7 +1,8 @@
-package io.th0rgal.oraxen.api.events;
+package io.th0rgal.oraxen.api.events.noteblock;
 
-import io.th0rgal.oraxen.mechanics.provided.gameplay.stringblock.StringBlockMechanic;
+import io.th0rgal.oraxen.mechanics.provided.gameplay.noteblock.NoteBlockMechanic;
 import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
@@ -11,35 +12,37 @@ import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
+public class OraxenNoteBlockInteractEvent extends Event implements Cancellable {
 
-    private final StringBlockMechanic mechanic;
-    private final Player player;
+    private final NoteBlockMechanic mechanic;
     private final Block block;
+    private final Player player;
     private final ItemStack itemInHand;
+    private final BlockFace blockFace;
     private final EquipmentSlot hand;
     private boolean isCancelled;
     private static final HandlerList HANDLERS = new HandlerList();
 
-    public OraxenStringBlockPlaceEvent(@NotNull final StringBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player, @Nullable final ItemStack itemInHand, @NotNull final EquipmentSlot hand) {
+    public OraxenNoteBlockInteractEvent(@NotNull final NoteBlockMechanic mechanic, @NotNull final Player player, @Nullable final ItemStack itemInHand, @NotNull final EquipmentSlot hand, @NotNull final Block block, @NotNull final BlockFace blockFace) {
         this.mechanic = mechanic;
-        this.player = player;
         this.block = block;
+        this.player = player;
         this.itemInHand = itemInHand;
+        this.blockFace = blockFace;
         this.hand = hand;
         this.isCancelled = false;
     }
 
     /**
-     * @return The StringBlockMechanic of this block
+     * @return The NoteBlockMechanic of this block
      */
     @NotNull
-    public StringBlockMechanic getMechanic() {
+    public NoteBlockMechanic getMechanic() {
         return mechanic;
     }
 
     /**
-     * @return The player who broke this block, or null if the event is not triggered by player
+     * @return The player who interacted with this block
      */
     @NotNull
     public Player getPlayer() {
@@ -47,7 +50,7 @@ public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
     }
 
     /**
-     * @return The block that was placed
+     * @return The block that was interacted with
      */
     @NotNull
     public Block getBlock() {
@@ -55,10 +58,15 @@ public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
     }
 
     /**
-     * Gets the item in the player's hand when they placed the furniture.
-     *
-     * @return The ItemStack for the item in the player's hand when they
-     *     placed the furniture
+     * @return The BlockFace that was clicked
+     */
+    @NotNull
+    public BlockFace getBlockFace() {
+        return blockFace;
+    }
+
+    /**
+     * @return The item in hand when the player interacted with the note block
      */
     @Nullable
     public ItemStack getItemInHand() {
@@ -66,9 +74,7 @@ public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
     }
 
     /**
-     * Gets the hand which the item was placed with.
-     *
-     * @return The hand used to place the item
+     * @return The hand used to perform interaction
      */
     @NotNull
     public EquipmentSlot getHand() {
@@ -85,6 +91,7 @@ public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
         isCancelled = cancel;
     }
 
+
     @NotNull
     @Override
     public HandlerList getHandlers() {
@@ -96,4 +103,3 @@ public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
     }
 
 }
-
