+++ src/main/java/io/th0rgal/oraxen/api/events/noteblock/OraxenNoteBlockPlaceEvent.java
@@ -1,8 +1,7 @@
-package io.th0rgal.oraxen.api.events;
+package io.th0rgal.oraxen.api.events.noteblock;
 
-import io.th0rgal.oraxen.mechanics.provided.gameplay.stringblock.StringBlockMechanic;
+import io.th0rgal.oraxen.mechanics.provided.gameplay.noteblock.NoteBlockMechanic;
 import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
@@ -10,39 +9,36 @@ import org.bukkit.event.HandlerList;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-public class OraxenStringBlockInteractEvent extends Event implements Cancellable {
+public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
 
-    private final StringBlockMechanic mechanic;
+    private final NoteBlockMechanic mechanic;
     private final Player player;
     private final Block block;
     private final ItemStack itemInHand;
     private final EquipmentSlot hand;
-    private final BlockFace blockFace;
     private boolean isCancelled;
     private static final HandlerList HANDLERS = new HandlerList();
 
-    public OraxenStringBlockInteractEvent(@NotNull final StringBlockMechanic mechanic, @NotNull final Player player, @Nullable final ItemStack itemInHand, @NotNull final EquipmentSlot hand, @NotNull final Block block, @NotNull final BlockFace blockFace) {
+    public OraxenNoteBlockPlaceEvent(@NotNull final NoteBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player, @NotNull final ItemStack itemInHand, @NotNull final EquipmentSlot hand) {
         this.mechanic = mechanic;
-        this.itemInHand = itemInHand;
-        this.block = block;
         this.player = player;
-        this.isCancelled = false;
+        this.block = block;
+        this.itemInHand = itemInHand;
         this.hand = hand;
-        this.blockFace = blockFace;
+        this.isCancelled = false;
     }
 
     /**
-     * @return The StringBlockMechanic of this block
+     * @return The NoteBlockMechanic of this block
      */
     @NotNull
-    public StringBlockMechanic getMechanic() {
+    public NoteBlockMechanic getMechanic() {
         return mechanic;
     }
 
     /**
-     * @return The player who interacted with this StringBlock
+     * @return The player who broke this block
      */
     @NotNull
     public Player getPlayer() {
@@ -50,7 +46,7 @@ public class OraxenStringBlockInteractEvent extends Event implements Cancellable
     }
 
     /**
-     * @return The StringBlock that was interacted with
+     * @return The block that was broken
      */
     @NotNull
     public Block getBlock() {
@@ -58,23 +54,20 @@ public class OraxenStringBlockInteractEvent extends Event implements Cancellable
     }
 
     /**
-     * @return Clicked block face
+     * Gets the item in the player's hand when they placed the furniture.
+     *
+     * @return The ItemStack for the item in the player's hand when they
+     *     placed the furniture
      */
     @NotNull
-    public BlockFace getBlockFace() {
-        return blockFace;
-    }
-
-    /**
-     * @return The item in hand when the player interacted with the string block
-     */
-    @Nullable
     public ItemStack getItemInHand() {
         return itemInHand;
     }
 
     /**
-     * @return The hand used to perform interaction
+     * Gets the hand used to place the furniture.
+     *
+     * @return The EquipmentSlot for the hand used to place the furniture
      */
     @NotNull
     public EquipmentSlot getHand() {
