+++ src/main/java/io/th0rgal/oraxen/api/events/stringblock/OraxenStringBlockBreakEvent.java
@@ -1,6 +1,6 @@
-package io.th0rgal.oraxen.api.events;
+package io.th0rgal.oraxen.api.events.stringblock;
 
-import io.th0rgal.oraxen.mechanics.provided.gameplay.noteblock.NoteBlockMechanic;
+import io.th0rgal.oraxen.mechanics.provided.gameplay.stringblock.StringBlockMechanic;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
@@ -8,25 +8,15 @@ import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
 
-/**
- * Event fired right before a player damages a NoteBlock.
- * If cancelled, the block will not be damaged.
- * @see NoteBlockMechanic
- */
-public class OraxenNoteBlockDamageEvent extends Event implements Cancellable {
+public class OraxenStringBlockBreakEvent extends Event implements Cancellable {
 
-    private final NoteBlockMechanic mechanic;
+    private final StringBlockMechanic mechanic;
     private final Player player;
     private final Block block;
     private boolean isCancelled;
     private static final HandlerList HANDLERS = new HandlerList();
 
-    /**
-     * @param mechanic The NoteBlockMechanic of this block
-     * @param block The block that was damaged
-     * @param player The player who damaged this block
-     */
-    public OraxenNoteBlockDamageEvent(@NotNull final NoteBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player) {
+    public OraxenStringBlockBreakEvent(@NotNull final StringBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player) {
         this.mechanic = mechanic;
         this.player = player;
         this.block = block;
@@ -34,15 +24,15 @@ public class OraxenNoteBlockDamageEvent extends Event implements Cancellable {
     }
 
     /**
-     * @return The note block mechanic
+     * @return The StringBlockMechanic of this block
      */
     @NotNull
-    public NoteBlockMechanic getMechanic() {
+    public StringBlockMechanic getMechanic() {
         return mechanic;
     }
 
     /**
-     * @return The player who broke the note block
+     * @return The player who broke this block
      */
     @NotNull
     public Player getPlayer() {
@@ -76,4 +66,5 @@ public class OraxenNoteBlockDamageEvent extends Event implements Cancellable {
     public static HandlerList getHandlerList() {
         return HANDLERS;
     }
+
 }
