+++ src/main/java/io/th0rgal/oraxen/api/events/stringblock/OraxenStringBlockDamageEvent.java
@@ -1,6 +1,6 @@
-package io.th0rgal.oraxen.api.events;
+package io.th0rgal.oraxen.api.events.stringblock;
 
-import io.th0rgal.oraxen.mechanics.provided.gameplay.noteblock.NoteBlockMechanic;
+import io.th0rgal.oraxen.mechanics.provided.gameplay.stringblock.StringBlockMechanic;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
@@ -8,15 +8,25 @@ import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
 
-public class OraxenNoteBlockBreakEvent extends Event implements Cancellable {
+/**
+ * Event fired right before a player damages a StringBlock
+ * If cancelled, the block will not be damaged.
+ * @see StringBlockMechanic
+ */
+public class OraxenStringBlockDamageEvent extends Event implements Cancellable {
 
-    private final NoteBlockMechanic mechanic;
+    private final StringBlockMechanic mechanic;
     private final Player player;
     private final Block block;
     private boolean isCancelled;
     private static final HandlerList HANDLERS = new HandlerList();
 
-    public OraxenNoteBlockBreakEvent(@NotNull final NoteBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player) {
+    /**
+     * @param mechanic The StringBlockMechanic of this block
+     * @param block The block that was damaged
+     * @param player The player who damaged this block
+     */
+    public OraxenStringBlockDamageEvent(@NotNull final StringBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player) {
         this.mechanic = mechanic;
         this.block = block;
         this.player = player;
@@ -24,15 +34,15 @@ public class OraxenNoteBlockBreakEvent extends Event implements Cancellable {
     }
 
     /**
-     * @return The NoteBlockMechanic of this block
+     * @return The StringBlockMechanic of the damaged block
      */
     @NotNull
-    public NoteBlockMechanic getMechanic() {
+    public StringBlockMechanic getMechanic() {
         return mechanic;
     }
 
     /**
-     * @return The player who broke this block
+     * @return The player who damaged the StringBlock
      */
     @NotNull
     public Player getPlayer() {
@@ -40,7 +50,7 @@ public class OraxenNoteBlockBreakEvent extends Event implements Cancellable {
     }
 
     /**
-     * @return The block that was broken
+     * @return The block that was damaged
      */
     @NotNull
     public Block getBlock() {
@@ -66,5 +76,4 @@ public class OraxenNoteBlockBreakEvent extends Event implements Cancellable {
     public static HandlerList getHandlerList() {
         return HANDLERS;
     }
-
 }
