+++ src/main/java/io/th0rgal/oraxen/api/events/stringblock/OraxenStringBlockPlaceEvent.java
@@ -1,6 +1,6 @@
-package io.th0rgal.oraxen.api.events;
+package io.th0rgal.oraxen.api.events.stringblock;
 
-import io.th0rgal.oraxen.mechanics.provided.gameplay.noteblock.NoteBlockMechanic;
+import io.th0rgal.oraxen.mechanics.provided.gameplay.stringblock.StringBlockMechanic;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
@@ -9,10 +9,11 @@ import org.bukkit.event.HandlerList;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
+public class OraxenStringBlockPlaceEvent extends Event implements Cancellable {
 
-    private final NoteBlockMechanic mechanic;
+    private final StringBlockMechanic mechanic;
     private final Player player;
     private final Block block;
     private final ItemStack itemInHand;
@@ -20,7 +21,7 @@ public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
     private boolean isCancelled;
     private static final HandlerList HANDLERS = new HandlerList();
 
-    public OraxenNoteBlockPlaceEvent(@NotNull final NoteBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player, @NotNull final ItemStack itemInHand, @NotNull final EquipmentSlot hand) {
+    public OraxenStringBlockPlaceEvent(@NotNull final StringBlockMechanic mechanic, @NotNull final Block block, @NotNull final Player player, @Nullable final ItemStack itemInHand, @NotNull final EquipmentSlot hand) {
         this.mechanic = mechanic;
         this.player = player;
         this.block = block;
@@ -30,15 +31,15 @@ public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
     }
 
     /**
-     * @return The NoteBlockMechanic of this block
+     * @return The StringBlockMechanic of this block
      */
     @NotNull
-    public NoteBlockMechanic getMechanic() {
+    public StringBlockMechanic getMechanic() {
         return mechanic;
     }
 
     /**
-     * @return The player who broke this block
+     * @return The player who broke this block, or null if the event is not triggered by player
      */
     @NotNull
     public Player getPlayer() {
@@ -46,7 +47,7 @@ public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
     }
 
     /**
-     * @return The block that was broken
+     * @return The block that was placed
      */
     @NotNull
     public Block getBlock() {
@@ -59,15 +60,15 @@ public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
      * @return The ItemStack for the item in the player's hand when they
      *     placed the furniture
      */
-    @NotNull
+    @Nullable
     public ItemStack getItemInHand() {
         return itemInHand;
     }
 
     /**
-     * Gets the hand used to place the furniture.
+     * Gets the hand which the item was placed with.
      *
-     * @return The EquipmentSlot for the hand used to place the furniture
+     * @return The hand used to place the item
      */
     @NotNull
     public EquipmentSlot getHand() {
@@ -95,3 +96,4 @@ public class OraxenNoteBlockPlaceEvent extends Event implements Cancellable {
     }
 
 }
+
