+++ src/main/java/io/th0rgal/oraxen/utils/breaker/BreakerSystem.java
@@ -12,9 +12,9 @@ import com.comphenix.protocol.wrappers.EnumWrappers;
 import io.th0rgal.oraxen.OraxenPlugin;
 import io.th0rgal.oraxen.api.OraxenBlocks;
 import io.th0rgal.oraxen.api.OraxenFurniture;
-import io.th0rgal.oraxen.api.events.OraxenFurnitureDamageEvent;
-import io.th0rgal.oraxen.api.events.OraxenNoteBlockDamageEvent;
-import io.th0rgal.oraxen.api.events.OraxenStringBlockDamageEvent;
+import io.th0rgal.oraxen.api.events.furniture.OraxenFurnitureDamageEvent;
+import io.th0rgal.oraxen.api.events.noteblock.OraxenNoteBlockDamageEvent;
+import io.th0rgal.oraxen.api.events.stringblock.OraxenStringBlockDamageEvent;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.block.BlockMechanic;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.furniture.FurnitureMechanic;
 import io.th0rgal.oraxen.mechanics.provided.gameplay.noteblock.NoteBlockMechanic;
